pyproject.toml
[project]
name = "langgraph-trip-planner"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.10.1",
    "ddgs>=9.5.5",
    "dotenv>=0.9.9",
    "hf-transfer>=0.1.9",
    "langchain-huggingface>=0.3.1",
    "langchain-openai>=0.3.32",
    "langgraph>=0.6.6",
    "sentencepiece>=0.2.1",
    "streamlit>=1.49.1",
    "tavily-python>=0.7.11",
    "torch>=2.8.0",
    "transformers>=4.56.1",
]
